adding user validation on summary delete

+++ src/Summaries/Aplication/UseCases/DeleteSummary.ts
@@ -1,4 +1,5 @@
 import { ValidateSchema } from "../../../Shared/Application/ValidateSchema";
+import { AppError } from "../../../Shared/Interface/Responses/AppError";
 import { IdSchema } from "../../../Users/Interface/Schemas/IdSchema";
 import { SummaryRepo } from "../../Domain/Repositories/SummaryRepo";
 import { DocumentRepository } from "../../Infrastructure/Repositories/CloudfareRepositoryR2";
@@ -15,7 +16,11 @@ export class DeleteSummaryUseCase {
     this.DocumentRepository = documentRepo;
     this.idValidate = new ValidateSchema(IdSchema);
   }
-  async exec(id: string) {
+  async exec(id: string,userId:string) {
+     const summary = await this.SummaryRepository.findById(id)
+    if (summary?.getAuthor() != userId){
+        throw new AppError("Only the owner can delete",403)
+    }
     this.idValidate.validate(id);
     await this.DocumentRepository.delete(id);
     await this.SummaryRepository.delete(id);
